Move suggestion button to topbar folder
